added add to cart feature, remove from cart feature, and updated cart page
