+++ NUShop/NUShop/NUShop/urls.py
@@ -26,5 +26,6 @@ urlpatterns = [
     path('products/', include('product.urls')),
     path('dashboard/', include('dashboard.urls')),
     path('inbox/', include('chat.urls')),
+    path('checkout/', include('checkout.urls')),
     path('admin/', admin.site.urls),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
