+++ NUShop/NUShop/checkout/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Cart, CartProduct
+
+admin.site.register(Cart)
+admin.site.register(CartProduct)
