+++ NUShop/NUShop/checkout/templates/checkout/index.html
@@ -1,0 +1,75 @@
+{% extends 'core/base.html' %}
+
+{% block title %}Cart{% endblock %}
+
+{% block content %}
+    <div class="grid grid-cols-6 gap-4">
+        <div class="col-span-4">
+            <div class="list">
+                    {% for product in products %}
+                        <div class="p-3 rounded-xl mb-3 mt-3 border-nushopprimary">
+                            <a href="{% url 'product:detail' product.product.id %}">
+                                <div class="grid grid-rows-5">
+                                    <div class="row-span-1">
+                                        <div class="grid grid-cols-10 gap-2">
+                                            <div class="col-span-1">
+                                                <input type="checkbox" name="selected_products" value="{{ cart_product.pk }}"> 
+                                            </div>
+                                            <div class="col-span-7 text-2xl">
+                                                <h1>{{ product.product.created_by }}</h1>
+                                            </div>
+                                            <div class="col-span-2">
+                                                <button class="mt-2 px-2 text-xl bt-nushoptertiary-2 rounded-xl text-center">View Shop</button>
+                                            </div>
+                                        </div>
+                                    </div>
+
+                                    <div class="row-span-3">
+                                        <div class="grid grid-cols-5 gap-2">
+                                            <div class="col-span-1">
+                                                <img src="{{ product.product.image.url }}" class="rounded-xl img-size">
+                                            </div>
+                                            <div class="col-span-3">
+                                                <h2 class="text-3xl text-black">{{ product.product.name }}</h2>
+                                                <p class="text-black text-xl">Variation:</p>
+                                            </div>
+                                            <div class="col-span-1">
+                                                <p class="text-3xl text-yellow-500">$ {{ product.product.price }}</p></p>
+                                            </div>
+                                        </div>
+                                    </div>
+
+                                    <div class="row-span-1">
+                                        <div class="grid grid-cols-5 gap-2">
+                                            <div class="col-span-1"></div>
+                                            <div class="col-span-2">
+                                                <p class="text-xl">Quantity: {{ product.quantity }}</p>
+                                            </div>
+                                            <div class="col-span-2">
+                                                <a href="{% url 'checkout:remove_from_cart' product.product.pk %}" class="mt-2 px-6 text-xl bt-nushoptertiary-2 rounded-xl text-center">Remove from Cart</a>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </a>
+                        </div>
+                    {% endfor %}
+            </div>
+        </div>
+        <div class="col-span-2 border-nushopprimary-2 rounded-xl m-12">
+            <div class="text-white text-3xl text-center bg-nushopprimary p-3 rounded-t-lg">
+                <h1>Summary</h1>
+            </div>
+            <div class="text-left text-black m-6">
+                <div class="text-xl">
+                    <p>Subtotal:</p>
+                    <p>Shipping Fee:</p>
+                </div>
+                <p class="text-3xl">Total:</p>
+            </div>
+            <div class="m-6">
+                <button class="text-xl bt-nushopsecondary rounded-xl text-center">Checkout</button>
+            </div>
+        </div>
+    </div>
+{% endblock %}
