+++ NUShop/NUShop/core/templates/core/base.html
@@ -27,10 +27,22 @@
               <a href="{% url 'product:products' %}" class="text-2xl cinzel text-nushopprimary hover:text-yellow-500">Browse</a>
               <a href="{% url 'chat:inbox' %}" class="text-2xl cinzel text-nushopprimary hover:text-yellow-500">Chat</a>
               <a href="{% url 'product:new' %}" class="text-2xl cinzel text-nushopprimary hover:text-yellow-500">Add Product</a>
-              <a href="{% url 'dashboard:cart' %}" class="text-2xl cinzel text-nushopprimary hover:text-yellow-500">Cart</a>
+              <a href="{% url 'checkout:index' %}" class="text-2xl cinzel text-nushopprimary hover:text-yellow-500">Cart</a>
               <a href="{% url 'dashboard:index' %}" class="text-2xl cinzel text-nushopprimary hover:text-yellow-500">Account</a>
           </div>
         </nav>
+        <div>
+            {% if messages %}
+                {% for message in messages %}
+                <div class="alert alert-success alert-dismissible fade show" role="alert">
+                  {{ message }}
+                  <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                  </button>
+                </div>
+                {% endfor %}
+            {% endif %}
+        </div>
         <div class="px-6 pb-6">
             {% block content %}
             {% endblock %}
