--- NUShop/NUShop/dashboard/templates/dashboard/cart.html
@@ -1,54 +1,0 @@
-{% extends 'core/base.html' %}
-
-{% block title %}Cart{% endblock %}
-
-{% block content %}
-    <div class="grid grid-cols-6 gap-4">
-        <div class="col-span-4">
-            <div class="list">
-                {% for product in products %}
-                    <div class="p-3 rounded-xl mb-3 mt-3 border-nushopprimary-1">
-                        <a href="{% url 'product:detail' product.id %}">
-                            <div class="grid grid-cols-5 gap-2">
-                                <div class="col-span-4 text-2xl">
-                                    <h1>NUS Computing</h1>
-                                </div>
-                                <div class="col-span-1">
-                                    <button class="mt-2 text-xl bt-nushoptertiary-2 rounded-xl text-center">View Shop</button>
-                                </div>
-                                <div class="col-span-1">
-                                    <img src="{{ product.image.url }}" class="rounded-xl img-size">
-                                </div>
-                                <div class="col-span-3">
-                                    <h2 class="text-3xl text-black">{{ product.name }}</h2>
-                                    <p class="text-2xl">Quantity:</p>
-                                    <p class="text-black text-2xl">Variation:</p>
-                                </div>
-                                <div class="col-span-1">
-                                    <p class="text-3xl text-yellow-500">$ {{ product.price }}</p></p>
-                                </div>        
-                            </div>
-                        </a>
-                    </div>
-                {% endfor %}
-            </div>
-        </div>
-        <div class="col-span-2 border-nushopprimary-3 rounded-xl m-12">
-            <div class="text-white text-3xl text-center bg-nushopprimary p-3 rounded-t-lg">
-                <h1>Summary</h1>
-            </div>
-            <div class="text-left text-black m-6">
-                <div class="text-xl">
-                    <p>Subtotal:</p>
-                    <p>Shipping Fee:</p>
-                </div>
-                <p class="text-3xl">Total:</p>
-            </div>
-            <div class="m-6">
-                <a href="{% url 'dashboard:payment' %}">
-                    <button class="text-xl bt-nushopsecondary rounded-xl text-center">Proceed to Pay</button>
-                </a>
-            </div>
-        </div>
-    </div>
-{% endblock %}
