+++ NUShop/NUShop/dashboard/templates/dashboard/view-seller.html
@@ -7,7 +7,7 @@
 {% endblock %}  
 
 {% block content %}
-    <div class="mt-6 px-3 py-6 border-nushopprimary rounded-xl">
+    <div class="mt-6 px-3 py-6 border-nushopprimary-2 rounded-xl">
         <div class="grid grid-cols-7">
             <div class="w-full col-span-1 overflow-hidden center">
                 {% if seller.image %}
