+++ NUShop/NUShop/product/templates/product/detail.html
@@ -86,6 +86,10 @@
                 </ul>
             </div>
             <p>Quantity</p>
+            <div class="mt-4">
+                <a href="{% url 'checkout:add_to_cart' product.id %}" class="mt-2 px-6 text-xl bt-nushoptertiary-2 rounded-xl text-center">Add to Cart</a>
+                <a href="#" class="mt-2 px-6 text-xl bt-nushoptertiary-1 rounded-xl text-center">Buy</a>
+            </div>
         </div>
     </div>
     <div class="col-span-2"></div>
